postmarketOS: detect if on systemd and use systemd
Give users the option to use systemd on postmarketOS.

Follow source code for `systemd-notify --booted` which is the standard
way to check if the system is using systemd.

https://www.freedesktop.org/software/systemd/man/latest/sd_booted.html#Notes

https://github.com/systemd/systemd/blob/52e9bc2deb3bec07780ea39a827809a68ffbd329/src/libsystemd/sd-daemon/sd-daemon.c#L757

+++ configs/default_hooks/sxmo_hook_restart_modem_daemons.sh
@@ -18,15 +18,12 @@ daemon_start() {
 		sxmo_notify_user.sh "$1 is already running"
 		return 0
 	fi
-	case "$SXMO_OS" in
-		alpine|postmarketos)
-			doas rc-service "$1" start
-			;;
-		arch|archarm|debian|nixos)
-			[ "$1" = "modemmanager" ] && set -- ModemManager
-			doas systemctl start "$1"
-			;;
-	esac
+	if [ -d /run/systemd/system ]; then
+		[ "$1" = "modemmanager" ] && set -- ModemManager
+		doas systemctl start "$1"
+	else
+		doas rc-service "$1" start
+	fi
 }
 
 daemon_stop() {
@@ -38,15 +35,12 @@ daemon_stop() {
 		sxmo_notify_user.sh "$1 is already stopped"
 		return 0
 	fi
-	case "$SXMO_OS" in
-		alpine|postmarketos)
-			doas rc-service "$1" stop
-			;;
-		arch|archarm|nixos)
-			[ "$1" = "modemmanager" ] && set -- ModemManager
-			doas systemctl stop "$1"
-			;;
-	esac
+	if [ -d /run/systemd/system ]; then
+		[ "$1" = "modemmanager" ] && set -- ModemManager
+		doas systemctl stop "$1"
+	else
+		doas rc-service "$1" stop
+	fi
 }
 
 daemon_isrunning() {
@@ -54,28 +48,23 @@ daemon_isrunning() {
 		sxmo_log "$1 does not exist on the system"
 		return 1
 	fi
-	case "$SXMO_OS" in
-		alpine|postmarketos)
-			rc-service "$1" status | grep -q started
-			;;
-		arch|archarm|debian|nixos)
-			[ "$1" = "modemmanager" ] && set -- ModemManager
-			systemctl is-active --quiet "$1"
-			;;
-	esac
+	if [ -d /run/systemd/system ]; then
+		[ "$1" = "modemmanager" ] && set -- ModemManager
+		systemctl is-active --quiet "$1"
+	else
+		rc-service "$1" status | grep -q started
+	fi
 }
 
 daemon_exists() {
-	case "$SXMO_OS" in
-		alpine|postmarketos)
-			[ -f /etc/init.d/"$1" ]
-			;;
-		arch|archarm|debian|nixos)
-			[ "$1" = "modemmanager" ] && set -- ModemManager
-			systemctl status "$1" > /dev/null 2>&1
-			[ $? -ne 4 ]
-			;;
-	esac
+
+	if [ -d /run/systemd/system ]; then
+		[ "$1" = "modemmanager" ] && set -- ModemManager
+		systemctl status "$1" > /dev/null 2>&1
+		[ $? -ne 4 ]
+	else
+		[ -f /etc/init.d/"$1" ]
+	fi
 }
 
 if command -v eg25-manager > /dev/null; then
